pyproject.toml
[project]
name = "db-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp",
    "psycopg[binary]",
    "python-dotenv",
]
